Refactor: reordering files
