+++ src/users/StudentCli.java
@@ -1,27 +1,28 @@
-package cli.users;
+package users;
 
-import cli.UserManagerCli;
 import attendance.AttendanceCli;
-import cli.BatchCli;
+import batch.BatchCli;
 import cli.models.CLI;
 import cli.models.menu.AuthMenu;
 import cli.models.menu.Menu;
 import config.enums.Role;
-import users.UserManager;
 
-public class StaffCli extends CLI {
-    private final AttendanceCli attendanceCli = new AttendanceCli(Role.STAFF);
-    private final BatchCli batchCli  = new BatchCli(Role.STAFF);
 
-    public StaffCli() {
+public class StudentCli extends CLI {
+    private final AttendanceCli attendanceCli = new AttendanceCli(Role.STUDENT);
+    private final BatchCli batchCli  = new BatchCli(Role.STUDENT);
+
+
+    public StudentCli() {
         super();
         addOptions();
     }
 
     @Override
     protected Menu createMenu() {
-        return  new AuthMenu();
+        return new AuthMenu();
     }
+
 
     @Override
     protected void addOptions() {
